requirements.txt
RPi.GPIO
apscheduler
